Move the I18N_LOCALE_CHANGE logic inside the useSetLocale hook
