+++ packages/ra-core/src/i18n/useSetLocale.tsx
@@ -1,6 +1,9 @@
-import { useContext } from 'react';
+import { useContext, useCallback } from 'react';
 
 import { TranslationContext } from './TranslationContext';
+import { I18N_CHANGE_LOCALE } from '../types';
+import { useUpdateLoading } from '../loading';
+import { useNotify } from '../sideEffect';
 
 /**
  * Set the current locale using the TranslationContext
@@ -30,9 +33,31 @@ import { TranslationContext } from './TranslationContext';
  *     );
  * }
  */
-const useSetLocale = () => {
-    const { setLocale } = useContext(TranslationContext);
-    return setLocale;
+const useSetLocale = (): SetLocale => {
+    const { setLocale, i18nProvider } = useContext(TranslationContext);
+    const { startLoading, stopLoading } = useUpdateLoading();
+    const notify = useNotify();
+    return useCallback(
+        (newLocale: string) =>
+            new Promise(resolve => {
+                startLoading();
+                // so we systematically return a Promise for the messages
+                // i18nProvider may return a Promise for language changes,
+                resolve(i18nProvider(I18N_CHANGE_LOCALE, newLocale));
+            })
+                .then(() => {
+                    stopLoading();
+                    setLocale(newLocale);
+                })
+                .catch(error => {
+                    stopLoading();
+                    notify('ra.notification.i18n_error', 'warning');
+                    console.error(error);
+                }),
+        [i18nProvider, notify, setLocale, startLoading, stopLoading]
+    );
 };
+
+type SetLocale = (locale: String) => Promise<void>;
 
 export default useSetLocale;
